PIA-884: Move Equatable extensions to tvOS target
