+++ PIA VPN-tvOS/Login/Domain/Entities/LoginError.swift
@@ -16,3 +16,18 @@ enum LoginError: Error {
     case passwordWrongFormat
     case generic(message: String?)
 }
+
+extension LoginError: Equatable {
+    public static func == (lhs: LoginError, rhs: LoginError) -> Bool {
+        switch (lhs, rhs) {
+            case (.unauthorized, .unauthorized), (.expired, .expired), (.usernameWrongFormat, .usernameWrongFormat), (.passwordWrongFormat, .passwordWrongFormat):
+                return true
+            case let (.throttled(lhsDelay), .throttled(rhsDelay)):
+                return lhsDelay == rhsDelay
+            case let (.generic(lhsMessage), .generic(rhsMessage)):
+                return lhsMessage == rhsMessage
+            default:
+                return false
+        }
+    }
+}
