+++ PIA VPN-tvOS/Login/Domain/Entities/UserAccount.swift
@@ -21,3 +21,10 @@ struct UserAccount {
         self.info = info
     }
 }
+
+extension UserAccount: Equatable {
+    public static func == (lhs: UserAccount, rhs: UserAccount) -> Bool {
+        lhs.credentials == rhs.credentials
+        && lhs.info == rhs.info
+    }
+}
